services/py1/requirements.txt
Flask==1.0.0
requests==2.23.0
PyJWT==1.7.1
cryptography
